+++ tests/test_events.py
@@ -95,13 +95,13 @@ class ResourceEventsTest(BaseWebTest, unittest.TestCase):
         # We created two records, for each of them we updated the ``last_edit_datr``
         # field, so we received two events.
         self.assertIsNone(events[0].impacted_records[0]["old"].get("status"))
-        self.assertIsNone(events[0].impacted_records[0]["old"].get("last_edit_date"))
+        self.assertIsNone(events[0].impacted_records[0]["old"].get("last_edit_at"))
         self.assertEqual(events[0].payload["user_id"], "plugin:kinto-signer")
         self.assertEqual(events[0].impacted_records[0]["new"]["status"],
                          "work-in-progress")
         self.assertIn("basicauth:", events[0].impacted_records[0]["new"]["last_edit_by"])
-        self.assertNotEqual(events[0].impacted_records[0]["new"]["last_edit_date"],
-                            events[-1].impacted_records[0]["new"]["last_edit_date"])
+        self.assertNotEqual(events[0].impacted_records[0]["new"]["last_edit_at"],
+                            events[-1].impacted_records[0]["new"]["last_edit_at"])
 
     def test_resource_changed_is_triggered_for_to_review(self):
         before = len(listener.received)